+++ src/lib/components/ui/ascii/Discord.svelte
@@ -1,0 +1,58 @@
+<pre
+  class="harry-glowing"
+  aria-hidden="true"
+  style="white-space: pre;
+	font-family: monospace;
+	font-size: 1px;
+	line-height: 1px;
+	letter-spacing: unset;
+	transform: unset;
+	overflow-y: hidden;">                                                                                          
+                                      ...........                            ...........                                      
+                                 .................                          ..................                                
+                             ......................                        ......:::::.::::::....                             
+                          ...............::::::::..............................:::::::::::::::::::...                         
+                       .............:::::::::::::::...........................:::::::::::::::::::::::...                      
+                     ..........:::::::::::::::::::::........................::::::;;;;;;;;;;;;;;;:::;;::..                    
+                   .........:::::::::::::::::::::::::::...............:::::::::;;;;;;;;;;;;;;;;;;;;;;;;;;::.                  
+                  ........:::::::::::::;;;;;;;;;;;::::::::::::::::::::::::::;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;:.                 
+                 ......::::::::::;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;++++++++++++++++;;;;;;;;;;                 
+                 ...:::::::::;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;+++++++++++++++++++++++++;;;;;:                
+                ...:::::::;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;++++++++++++++++++++++++++++++++++;;:               
+               ..::::::;;;;;;;;;;;;;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++;;              
+              ..::::;;;;;;;;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++;;:             
+              ..:::;;;;;;;;;;;+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++;;             
+             .::;;;;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++;;            
+            .::;;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++;            
+            .:;;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++xx++xxxxxxxxxx+xxx++++++++++++;           
+           .:;;;;;;;++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++xxxxxxxxxxxxxxxxxxxxxxxxx+++++++++++          
+          ..:;;;;;+++++++++++++++++++++++++++++++++++++++++++++++++++++++++xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx++++++++         
+          ::;;;;;+++++++++++++++++++++++++++xxxxxx+x+++++++++++++++++++xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx+x+++++         
+         .:;;;;+++++++++++++++++++++++xxxx++;;;;;+xxxxxxxx+++++++++xxxxxxxxxx++;;;++xxxxxxxxxxxxxxxxxxxxxxxxxxxxxx+++         
+         :;;;;++++++++++++++++++++xxxxxx+:.........:+xxxxxxxxxxxxxxxxxxxxxx;.........:xxxxxxxxxxxxxxxxxxxxxxxxxxxxx+++        
+        .:;;;++++++++++++++++xxxxxxxxxx;.............;xxxxxxxxxxxxxxxxxxx;.............;xXxxxxxxxxxxxxxxxxxxxxxxxxx+++        
+        :;;+++++++++++++xxxxxxxxxxxxxx;...............;xxxxxxxxxxxxxxxxx;...............;xXxxxxxxxxxxxxxxxxxxxxxxxxxxx+       
+        ;;++++++++++xxxxxxxxxxxxxxxxx;.................+xxxxxxxxxxxxxxx+.................+XXxxxxxxxxxxxxxxxxxxxxxxxxxxxx      
+       :;++++++++xxxxxxxxxxxxxxxxxxxx:.................;xXxxxxxxxxxxxxX:.................;XXXxxxxxxxxxxxxxxxxxxxxxxxxxxx      
+      :;++++++xxxxxxxxxxxxxxxxxxxxxXx:.................;xXxxxxxxxxxxxXX:.................;XXXXxxxxxxxxxxxxxxxxxxxxxxxxxx      
+      ;;++++xxxxxxxxxxxxxxxxxxxxxxxXX;.................+XXXxxxxxxxxxxXX;.................;XXXXXXxxxxxxxxxxxxxxxxxxxxxxxx      
+      +++++xxxxxxxxxxxxxxxxxxxxxxxXXXx:...............:xXXXXXxxxxxxXXXXx:...............;x$XXXXXXXXXxxxxxxxxxxxxxxxxxxxxx     
+      ++xxxxxxxxxxxxxxxxxxxxxxxxxXXXXXx:.............:xXXXXXXXXXXXXXXXXXx:.............:+XXXXXXXXXXXXXXXXXXXXXXxxxxxxxxxx     
+     ++xxxxxxxxxxxxxxxxxxxxxXXXXXXXXXXXX;:..........;x$XXXXXXXXXXXXXXXXXXX;...........;xXXXXXXXXXXXXXXXXXXXXXXXXXXXxXXxXX     
+     +xxxxxxxxxxxxXXXXXXXXXXXXXXXXXXXXXXXx+;:::::;+xXXXXXXXXXXXXXXXXXXXXXXXx+;:::::;+xXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX     
+     xxxxxxxxxxXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx    
+     xxxxxxXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXx    
+     xxXxXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX    
+     xXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX    
+     XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX    
+     XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX    
+     XXX$$XXXXXXXXXXXXXXXXXXXXX$$$$$XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$XXXXXXXXXXXXXXXXXXXX$$XX     
+      XXX$$$$$XXXXXXXXXXXXXXXX$$$$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$XXXXXXXXXXXXXXX$$$$$$XXX     
+        XXX$$$$$$$$XXXXXXXXXXXXX$$X $$$$$$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$X$$$$$XXXXXXXXXXX$$$$$$$$XXX       
+           xX$$$$$$$$$$$XXXXXXXX$$$++     X$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$     ;xX$$XXXXXXX$$$$$$$$$$$XXX          
+             XXX$$$$$$$$$$$$XXXXXX$$XXxx+;     XXXXXXXXX$XX$$$$$$$$$$$$XXXXXXXXx    ;+xXXX$$XXXXXX$$$$$$$$$$$$$XXx            
+                XXX$$$$$$$$$$$$XXXXXXX$$Xxx                                        +xX$XXXXXXXXX$$$$$$$$$$$$XXX               
+                    XX$$$$$$$$$$$$XXXXXXXXX                                        X$XXXXXXX$$$$$$$$$$$$XXx                   
+                       XXX$$$$$$$$$$$$$XXX                                          XX$$$$$$$$$$$$$$XXX                       
+                             XXXXXXXXXXX                                              XXXXXXXXXX$X                            
+</pre>
