+++ src/lib/components/ui/ascii/Email.svelte
@@ -1,0 +1,58 @@
+<pre
+  aria-hidden="true"
+  class="harry-glowing"
+  style="white-space: pre;
+	font-family: monospace;
+	font-size: 1px;
+	line-height: 1px;
+	letter-spacing: unset;
+	transform: unset;
+	overflow-y: hidden;">
+          xxx+xx;...........................................................................................;xXXXXXX          
+        +xxxxxxx+;.........................................................................................;xXXXXXXXXX        
+       xxxxxxxxxxxx+:...................................................................................:+XXXXXXXXXXXX        
+      xxxxxxxxxxxxxxx+;...............................................................................;xXXXXXXXXXXXXXX        
+      xxxxxxxxxxxxxxxxxx+:.........................................................................:+xXXXXXXXXXXXXXXXX        
+      xxxxxxxxxxxxxxxxxxxXx+.....................................................................;XXXXXXXXXXXXXXXXXXXXX       
+       xXxxxxxxxxxxxxxxxxxxXx+:...............................................................:+XXXXXXXXXXXXXXXXXXX$$$$       
+       $$XXxXXXXXXXXXXXXXXXXXXXx;...........................................................;xXXXXXXXXXXXXXXXXXXXXX$$$$       
+       $$$$$XXXXXXXXXXXXXXXXXXXXXXx:......................................................+XXXXXXXXXXXXXXXXXXXXXX$$$$$$       
+       $$$$$$$$XXXXXXXXXXXXXXXXXXXXXx;.................................................;xXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$       
+       $$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXXx;...........................................:+XXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$       
+       $$$$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXXX+.......................................;xX$XXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXXx;.................................:+XXXXXXXXXXXXXXXXXXX$$$XX$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$X+xXXXXXXXXXXXXXXXXXXXXXXX+:............................;XXXXXXXXXXXXXXX$$$$$$$XX+;+$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$x;;+xXXXXXXXXXXXXXXXXXXXXXXXx;.......................:+X$$XXXXXXX$X$$$$$$$$$$Xx+;::+$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$x:::;;+xXXXXXXXXXXXXXXXXXXXXXXX+:..................;xX$$XX$$$$$$$$$$$$$$$$XX+;;::::+$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$x:.::::;++XXXXXXXXXXXXXXXXXXXXXXXX;.............:xX$$X$$$$$$$$$$$$$$$$$$Xx;;::::::.+$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+....::::;;+xXXXXXXXXXXXXXXXXXXXX$$X+:........;xX$$$$$$$$$$$$$$$$$$$$XX+;:::::.....;$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+.......::::;;+XXX$XX$XXXX$$$$$$$$$$$XX;...:+X$$$$$$$$$$$$$$$$$$$$$Xx;;::::........;$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+.........::::;;+xX$$$$$$$$$$$$$$$$$$$$$$XX$$$$$$$$$$$$$$$$$$$$$$X+;::::...........;$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+............::::;;xXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$Xx;;:::..............;$$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..............::::;;+xX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X+;::::................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+.................::::;;xX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$Xx;;:::...................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+...................::::;;+xX$$$$$$$$$$$$$$$$$$$$$$$$$$X+;::::.....................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+......................::::;;+X$$$$$$$$$$$$$$$$$$$$$Xx;;:::........................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+........................::::;;+xX$$$$$$$$$$$$$$$$X+;::::..........................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+...........................::::;;+X$$$$$$$$$$$Xx;;:::.............................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+.............................:::::;+x$$$$$$$X+;::::...............................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+................................::::;;xX$Xx;::::..................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................::;;;;;:::::....................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................::::............................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+.................................:................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$$       
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$        
+       $$$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$        
+        $$$$$$$$$$$$$+..................................................................................;X$$$$$$$$$$$$        
+         $$$$$$$$$$$$+..................................................................................:X$$$$$$$$$$          
+            </pre>
