+++ src/lib/components/ui/ascii/Forum.svelte
@@ -1,0 +1,70 @@
+<pre
+  aria-hidden="true"
+  class="harry-glowing"
+  style="white-space: pre;
+	font-family: monospace;
+	font-size: 1px;
+	line-height: 1px;
+	letter-spacing: unset;
+	transform: unset;
+	overflow-y: hidden;">                                                                                          
+             ........................................:.....::::::::::::::::::::::::::::::::::::::::::::::::::::::             
+          ...:;+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx++;:           
+         .:;+xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX++          
+         ;xXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$XX$$$$$$$$$$$$$$$$$$$$$$$$$XX          
+         xXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$XXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$X$XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXX$XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$XX$X$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXX$X+::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::+X$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXX$+....................................................................+$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXX$$x....................................................................+$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXX$$x:..................................................................:+$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXX$X$$x:..................................................................:+$$$$$$$$$$$$$$$$$$X         
+         XX$$$XXXXXXXXXX$X$$x:..................................................................:x$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXX$$X$XXxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxXX$$$$$$$$$$$$$$$$$$$X         
+         XX$$$XXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$XXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$$XXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXXXXXXxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXXXXx:...................................................;$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXX$$x....................................................:X$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXX$XXXXXXXXXXXXX$$x:...................................................;$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XXXXXXXXXXXXXXXXXX$x:...................................................;$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$XXXXXXXXXXXXXX$x:...................................................;$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$XXXXXXXXXXXXX$$X+::::::::::::::::::::::::::::::::::::::::::::::::::;x$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$XXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$XXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$XXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$$XXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$X$$$$XXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$$X$XXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX         
+         XX$$$$$$$$$XXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$$$$X$$XX$$XXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$$$$$$$$$$$XXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$X$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X         
+         XX$$$$$$$$$$$XXXXXXXXXX$$$$$$$$$$$$$$$XXXXXXXXXXXXXXXXXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X          
+         XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX$$XXXXX$XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XX          
+          XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XXXX$$X$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$X           
+            $XX$$$$$$$$$$$$$$$$$$$$$$$$$$$$XXXXXXX$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$$XXXXXXXX$$$$$$$$$$$$$$$$$$$$$X             
+                                 $$$$$$$XXXXXXXXXXX$$$$$$$$$$$$$XX                                                            
+                                 X$$$$$$XXXXXXXXXX$$$$$$$$$$$$XX                                                              
+                                 X$$$$$XXXXXXXXX$$$$$$$$$$$XX                                                                 
+                                 X$$$$$XXXXXXXX$$$$$$$$$$$X                                                                   
+                                  X$$$XXXXXXXX$$$$$$$$$$XX                                                                    
+                                  X$$$XXXXXX$$$$$$$$$$X$                                                                      
+                                  X$$$XXXXXX$$$$$$$$XX                                                                        
+                                  XX$XXXXX$$$$$$$$XX                                                                          
+                                  XXXXXXXX$$$$$$XX                                                                            
+                                 xXXXXXX$$$$$$$XX                                                                             
+                                 xXXXX$$$$$$$XX                                                                               
+                                  XX$$$$$$$XX                                                                                 
+                                   X$$$$$$X                                                                                   
+                                     X$$$                                                                                     
+
+</pre>
