+++ src/lib/components/ui/ascii/HandShake.svelte
@@ -1,0 +1,57 @@
+<pre
+  aria-hidden="true"
+  class="harry-glowing"
+  style="white-space: pre;
+	font-family: monospace;
+	font-size: 3px;
+	line-height: 3px;
+	letter-spacing: unset;
+	transform: unset;
+	overflow-y: hidden;">
+            ;:::;;;+                        +;;::::::::::::::::;;                                                             
+          :::...:::;;++                +++;;::..................:::::;                                                        
+        ::..........::;;;;;;;;;;;;;;;;;+++;:.........................::::;                                                    
+      ::...............::::::::::::;;;;++;:..............................:::;;;;                         +++;;;;;             
+     :.........................::::;;;++;:....................................:::::::;               ++++;;:::...:.           
+    ::..........................::::;++;:............................................:::;+      ++;;;;;;::..........          
+   ::............................::;;+;:...............:::..............................::;;;;;;;;::.................:        
+  ::.............................::;+;...............:;;;::;;;;::...........................:::........................       
+  :..............................:;;:..............::;::.......:;;:::..................................................:      
+ ::.............................::+;..............::;:............::;:::................................................:     
+ ::..............................:+;:............::::................:;;::::............................................::    
+;::..............................:;;;:........:::::....................::;;;:::..........................................::   
+;:::..............................:;+;:::::::;;::.........................:;;;;:::::....................................::::  
+;;::::::............................::;;;;;;;::..............................:;;;;;;:::::...............................::::  
++;;:::::::::....................................................................:;++;;;;;::::.........................::::::; 
+++;;;;;;::::::::::::...............................................................:;+++;;;;::::.....................:::::::: 
+ ++++;;;;;;;;;::::::::................................................................:;++++;;;:::........................::::
+  x+++++++++;;;;;;;;::::.................................................................:;++++;;:::.....................:::;;
+     +xx++++++++++;;;;::::..................................................................:;+++;;;::::..........:::::::;;;; 
+        ++xx++++++++;;;::::....................................................................:;+++;;::::::::::::::::;;++++  
+           xxxxx+++++;;;;::::....................................................................:;+++;;;;::::::::;;+++++     
+             xxxxx+++++;;;:::::.........................................;;;:.......................:;+++++;;;;;;;++++         
+               xxxx+++++;;;;:::::.........................................::::::...................::;++++++++++++            
+                 xxxx+++++;;;;:::::.............................................::::...............::;;++xx+++++x             
+                  xxxx++++++;;;;:::::..............................................:;;;:...........:::;++xxxxxx               
+                    xxxx+++++;;;;::::::...............................................:;;+;;::....::::;;++xxxx                
+                      xxxx+++++;;;;:::::.................................................:;+++;;::::;;;;;++x                  
+                       xxxxx++++;;;;:::::......................::::::......................:;+++++;;;;;;;;++                  
+                         xxxx+;:::;;;;:::::........................:;++;:....................:;+x+++;;;;;;++                  
+                          ++;:.....:;;;:::::..........................:;++;::................::;+++++;;;;;+                   
+                         x++;:......:;+;:::::............................:;+++;:.............::;;++x+++++;                    
+                         xx++;:....::;+;;:::::........::...................:;+++;;:.........::;;++++x+++                      
+                         xxx++;;:::;;+++;;;++;;;::::::;++;;::.................:;;+++;::::::;;;;+++++                          
+                          xxxx++;;;;++x+++++;::...:::..:;;+++;;:.................:;+++++++++++++++++                          
+                          xxxxx++++++xxx++;:........::...:::;;;;;;::................:;++xx++++++++x                           
+                           xxxxx+++xxxx++;:.........:;;;;;:::..::;++;;:.............::;++xxxxxx+                              
+                             xxxxxxxxx+++;::.......:;+;;::::::...::;;++;;;::......::;;+++++                                   
+                                    x++++;;::.....:;+;:......:;++;;;;;::;;++;;;;;;;++++++++                                   
+                                    x++++;;;:::::;++;:.......:;;:.....:::..::;;++++++++++x                                    
+                                    ++++++;;;;;;;+++;........:;:........:.....::;+x+xx                                        
+                                     x++++++;;;+++++;:.......:;.........;::::::;;                                             
+                                        +++++++++++;;:::...:;;;........:++;;;;                                                
+                                               ++++;;::::::;++;:......:;++                                                    
+                                                +++;;:::::;;+++;::::::;;;                                                     
+                                                 x++;;;:;;;;x+++;;;;;;++                                                      
+                                                    ++;;;;    +++++++++                                                       
+</pre>
