+++ src/lib/components/ui/ascii/Name.svelte
@@ -1,0 +1,28 @@
+<p class="sr-only">The Gamblers Guild</p>
+<pre aria-hidden="true" class=" harry" tabindex="-1">
+ooooooooooooo oooo                  
+8'   888   `8 `888                  
+     888       888 .oo.    .ooooo.  
+     888       888P"Y88b  d88' `88b 
+     888       888   888  888ooo888 
+     888       888   888  888    .o 
+    o888o     o888o o888o `Y8bod8P' 
+</pre>
+<pre aria-hidden="true" class=" harry" tabindex="-1">
+        .oooooo.                                 .o8       oooo                               
+       d8P'  `Y8b                               "888       `888                               
+      888            .oooo.   ooo. .oo.  .oo.    888oooo.   888   .ooooo.  oooo d8b  .oooo.o  
+      888           `P  )88b  `888P"Y88bP"Y88b   d88' `88b  888  d88' `88b `888""8P d88(  "8  
+      888     ooooo  .oP"888   888   888   888   888   888  888  888ooo888  888     `"Y88b.   
+      `88.    .88'  d8(  888   888   888   888   888   888  888  888    .o  888     o.  )88b  
+       `Y8bood8P'   `Y888""8o o888o o888o o888o  `Y8bod8P' o888o `Y8bod8P' d888b    8""888P'  
+</pre>
+<pre aria-hidden="true" class=" harry" tabindex="-1">
+       .oooooo.                 o8o  oooo        .o8
+      d8P'  `Y8b                `"'  `888       "888
+     888           oooo  oooo  oooo   888   .oooo888
+     888           `888  `888  `888   888  d88' `888
+     888     ooooo  888   888   888   888  888   888
+     `88.    .88'   888   888   888   888  888   888
+      `Y8bood8P'    `V88V"V8P' o888o o888o `Y8bod88P
+</pre>
