+++ src/lib/components/ui/ascii/index.ts
@@ -1,0 +1,8 @@
+import Dice from "./Dice.svelte";
+import Discord from "./Discord.svelte";
+import Email from "./Email.svelte";
+import HandShake from "./HandShake.svelte";
+import Forum from "./Forum.svelte";
+import Name from "./Name.svelte";
+
+export { Dice, Discord, Email, HandShake, Forum, Name };
